Squish can now die
Keep in mind the planets projectile damage is 100 so he will die right away. Make sure if you want to test it, that you turn it to a lower number. I'd advised against it because I will probably update it in my change and that may or may not cause a merge issue.